+++ install.sh
@@ -50,6 +50,12 @@ if [ ! -f htdocs/app/etc/local.xml ] ; then
       sed -i -e s/MAGENTO_DB_NAME/${MAGENTO_DB_NAME}/g .modman/Ffuenf_TestSetup/app/etc/local.xml.phpunit
     fi
 
+    # use magedownload-cli for magento > 1.9.0.0
+    VERSION=`echo ${MAGENTO_VERSION} | sed -n 's/.*-\(.*\)/\1/p'`
+    if [ ! -f /tmp/magento-ce-${VERSION}.tar.gz ] ; then
+        tools/magedownload file magento-${VERSION}.tar.gz /tmp/${MAGENTO_VERSION}.tar.gz
+    fi
+
     tools/n98-magerun install \
       --dbHost="${MAGENTO_DB_HOST}" --dbUser="${MAGENTO_DB_USER}" --dbPass="${MAGENTO_DB_PASS}" --dbName="${MAGENTO_DB_NAME}" --dbPort="${MAGENTO_DB_PORT}" \
       --installSampleData=no \