+++ setup.sh
@@ -6,6 +6,8 @@ function cleanup {
   if [ -z $SKIP_CLEANUP ]; then
     echo "Removing build directory ${BUILDENV}"
     rm -rf "${BUILDENV}"
+    rm -rf ${WORKSPACE}/build
+    rm -f magento-ce-${MAGENTO_VERSION}.tar.gz
   fi
 }
 
@@ -34,50 +36,53 @@ chmod +x ${BUILDENV}/tools/modman
 curl -s -L http://files.magerun.net/n98-magerun-latest.phar -o ${BUILDENV}/tools/n98-magerun
 chmod +x ${BUILDENV}/tools/n98-magerun
 mkdir -p /usr/local/share/n98-magerun/modules/
-if [ ! -f /usr/local/share/n98-magerun/modules/yireo ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/yireo ] ; then
   git clone https://github.com/yireo/magerun-addons /usr/local/share/n98-magerun/modules/yireo
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/mpmd ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/mpmd ] ; then
   git clone https://github.com/aoepeople/mpmd /usr/local/share/n98-magerun/modules/mpmd
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-commands ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-commands ] ; then
   git clone https://github.com/degdigital/magerun-commands /usr/local/share/n98-magerun/modules/magerun-commands
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-addons ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-addons ] ; then
   git clone https://github.com/peterjaap/magerun-addons /usr/local/share/n98-magerun/modules/magerun-addons
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-creatuity ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-creatuity ] ; then
   git clone https://github.com/creatuity/magerun-creatuity /usr/local/share/n98-magerun/modules/magerun-creatuity
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-module-cache-benchmark ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-module-cache-benchmark ] ; then
   git clone https://github.com/cmuench/magerun-module-cache-benchmark /usr/local/share/n98-magerun/modules/magerun-module-cache-benchmark
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/cmuench-magerun-addons ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/cmuench-magerun-addons ] ; then
   git clone https://github.com/cmuench/cmuench-magerun-addons /usr/local/share/n98-magerun/modules/cmuench-magerun-addons
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/kalenjordan-magerun-addons ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/kalenjordan-magerun-addons ] ; then
   git clone https://github.com/kalenjordan/magerun-addons /usr/local/share/n98-magerun/modules/kalenjordan-magerun-addons
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/Webgriffe_Golive ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/Webgriffe_Golive ] ; then
   git clone https://github.com/aleron75/Webgriffe_Golive /usr/local/share/n98-magerun/modules/Webgriffe_Golive
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/ffuenf-download-remote-media ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/ffuenf-download-remote-media ] ; then
   git clone https://github.com/ffuenf/download-remote-media /usr/local/share/n98-magerun/modules/ffuenf-download-remote-media
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/sxmlsv ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/sxmlsv ] ; then
   git clone https://github.com/KamilBalwierz/sxmlsv /usr/local/share/n98-magerun/modules/sxmlsv
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-dataprofile ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-dataprofile ] ; then
   git clone https://github.com/marcoandreini/magerun-dataprofile /usr/local/share/n98-magerun/modules/magerun-dataprofile
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/Magerun-DBClean ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/Magerun-DBClean ] ; then
   git clone https://github.com/steverobbins/Magerun-DBClean /usr/local/share/n98-magerun/modules/Magerun-DBClean
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/magerun-modman ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/magerun-modman ] ; then
   git clone https://github.com/fruitcakestudio/magerun-modman /usr/local/share/n98-magerun/modules/magerun-modman
 fi
-if [ ! -f /usr/local/share/n98-magerun/modules/EAVCleaner ] ; then
+if [ ! -d /usr/local/share/n98-magerun/modules/EAVCleaner ] ; then
   git clone https://github.com/magento-hackathon/EAVCleaner /usr/local/share/n98-magerun/modules/EAVCleaner
+fi
+if [ ! -d /usr/local/share/n98-magerun/modules/magescan ] ; then
+  git clone https://github.com/steverobbins/magescan /usr/local/share/n98-magerun/modules/magescan
 fi
 curl -s -L https://getcomposer.org/composer.phar -o ${BUILDENV}/tools/composer
 chmod +x ${BUILDENV}/tools/composer
@@ -87,6 +92,22 @@ curl -s -L https://scrutinizer-ci.com/ocular.phar -o ${BUILDENV}/tools/ocular
 chmod +x ${BUILDENV}/tools/ocular
 curl -s -L https://raw.github.com/lehmannro/assert.sh/master/assert.sh -o ${BUILDENV}/tools/assert.sh
 chmod +x ${BUILDENV}/tools/assert.sh
+
+curl -s -L http://magedownload.steverobbins.com/download/latest/magedownload.phar -o ${BUILDENV}/tools/magedownload
+chmod +x ${BUILDENV}/tools/magedownload
+echo "
+user:
+    id: ${MAGEDOWNLOAD_ID}
+    token: ${MAGEDOWNLOAD_TOKEN}
+" > ~/.magedownload-cli.yaml
+if [ -d /home/vagrant ] ; then
+  cp ~/.magedownload-cli.yaml /home/vagrant/.magedownload-cli.yaml
+fi
+
+cp ${BUILDENV}/.n98-magerun.yaml ~/.n98-magerun.yaml
+if [ -d /home/vagrant ] ; then
+  cp ${BUILDENV}/.n98-magerun.yaml ~/.n98-magerun.yaml
+fi
 
 cp -rf "${WORKSPACE}" "${BUILDENV}/.modman/"
 ${BUILDENV}/install.sh